+++ src/utils/redis/redis.js
@@ -2,34 +2,27 @@ const redis = require('redis');
 const bluebird = require('bluebird');
 bluebird.promisifyAll(redis.RedisClient.prototype);
 bluebird.promisifyAll(redis.Multi.prototype);
-const data = require('./mockedDataSearch');
 
 const REDIS_PORT = process.env.REDIS_PORT || 6379;
 // const client = redis.createClient({host: 'redis'});
 const client = redis.createClient();
 
-client.on('error', (err) => {
-  console.log('Error ' + err);
-});
-
-const writeSearchToCache = (result, query) => {
+const writeSearchToCache = async (query, result) => {
   let filteredResult = result.filter(e => e.city.toLowerCase() === query.toLowerCase());
-  console.log('filteredResult', filteredResult);
-  client.setAsync(query, JSON.stringify(filteredResult));
+  await client.setAsync(query, JSON.stringify(filteredResult));
 };
 
-const getSearchResults = (query) => {
-  console.log('~~~~~~~~query', query);
-  return client.getAsync(query)
-    .then( (value) => {
-      return JSON.parse(value);
-    })
-    .catch( (err) => console.log('an error?', err));
+const getSearchResults = async (query) => {
+  let value = await client.getAsync(query);
+  return JSON.parse(value);
 };
 
 // writeSearchToCache(data.listings, 'seattle');
 // getSearchResults('seattle');
-
+const closeInstance = async (callback) => {
+  await client.quit(callback);
+};
 
 module.exports.getSearchResults = getSearchResults;
 module.exports.writeSearchToCache = writeSearchToCache;
+module.exports.closeInstance = closeInstance;
